# api.yaml
openapi: '3.1.0'
servers:
  - url: telalucis.org
info:
  contact:
    name: David
  version: '1.0.0'
  title: 'cititations api'
  description: backend for citations browsing and querying

paths:
  /books:
    get:
      tags: 
        - Books
      description: get list of books
      parameters:
       - name: offset
         in: query
         required: false
         schema:
            type: number
            minimum: 0
            maximum: 10000
       - name: length
         in: query
         required: false
         schema:
           type: number
           minimum: 1
           maximum: 100
       - name: query
         description: search title and author
         in: query
         required: false
         schema:
          type: string      
      summary: Search books
      responses:
        '200':
          description: books list
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      count:
                        type: number
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/BookSummary'
  /books/{bookId}:
    get:
      tags:
        - Books
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      summary: Get a book's summary, with its pages in order under pageOrder.
      responses:
        '200':
          description: book summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookSummary'
        '404':
          description: no book with this id
  /books/{bookId}/pages/{pageId}:
    get:
      tags:
        - Books
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
        - name: pageId
          description: a pageId from the book's pageOrder
          in: path
          required: true
          schema:
            type: number
            minimum: 1
      summary: Get one page of a book, with the citations in other books that point to it.
      responses:
        '200':
          description: book page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookPage'
        '404':
          description: no such book or page
  /books/{bookId}/pages/{pageId}/insights:
    get:
      tags:
        - Books
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
        - name: pageId
          description: a pageId from the book's pageOrder
          in: path
          required: true
          schema:
            type: number
            minimum: 1
      summary: >-
        What other books in the collection say about this page: an LLM summary of the passages
        that cite it (each citing page with two pages before and after), with this page as context.
      responses:
        '200':
          description: the insights
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageInsights'
        '204':
          description: nothing in the collection cites this page (no summary is made)
        '404':
          description: no such book or page
        '502':
          description: the OpenRouter request failed
        '503':
          description: insights are unavailable (OPENROUTER_KEY is not set)
  /books/{bookId}/citationsTo:
    get:
      tags:
        - Citations
      parameters:
        - name: offset
          in: query
          required: false
          schema:
             type: number
             minimum: 0
             maximum: 10000
        - name: length
          in: query
          required: false
          schema:
            type: number
            minimum: 1
            maximum: 100
        - name: bookId
          in: path
          required: true
          schema:
            type: number
      summary: Get other books that cite this one.
      responses:
        200:
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      count:
                        type: number
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Citation'
components:
  schemas:
    BookPage:
      type: object
      properties:
        printedPageNumber:
          type: string
        pageNumber:
          type: number
        bookId:
          type: string
        foreignCitations:
          type: array
          items:
            $ref: '#/components/schemas/PageCitation'
        blocks:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              html:
                type: string
              citations:
                type: array
                items:
                  $ref: '#/components/schemas/PageCitation'                                
    PageInsights:
      type: object
      required:
        - bookId
        - pageId
        - printedPageNumber
        - overview
        - sources
      properties:
        bookId:
          type: string
        pageId:
          type: number
        printedPageNumber:
          type: string
        overview:
          description: how the citing sources, taken together, treat this page
          type: string
        sources:
          description: one per page in another book that cites this page, with what it says
          type: array
          items:
            type: object
            required:
              - bookId
              - title
              - author
              - pageId
              - printedPageNumber
              - footnoteIdentifiers
              - summary
            properties:
              bookId:
                type: string
              title:
                type: string
              author:
                type: string
              pageId:
                description: the citing page (a pageId of that book)
                type: number
              printedPageNumber:
                type: string
              footnoteIdentifiers:
                description: the footnote markers on that page that cite this page
                type: array
                items:
                  type: string
              summary:
                description: what that author says about this page ('' if the model gave none)
                type: string
    PageCitation:
      description: a Citation with the HTML of the page its footnote is on
      allOf:
        - $ref: '#/components/schemas/Citation'
        - type: object
          required:
            - sourcePageText
          properties:
            sourcePageText:
              description: the citing page's blocks' HTML, in reading order, one per line, without running headers and footers and with only this citation's footnote
              type: string
    BookSummary:
      type: object
      required:
        - id
        - title
        - author
        - pageCount
        - citedByCount
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        url:
          type: string
        coverPhotoPath:
          description: cover image, relative to the frontend's root (e.g. covers/<id>.jpg)
          type: string
        pageCount:
          type: number
        citedByCount:
          description: number of citations in other books that reference this book
          type: number
        pageOrder:
          type: array
          items:
            type: object
            properties:
              pageId:
                type: number
              printedPageNumber:
                type: string
              citedByCount:
                description: number of citations in other books that reference this page
                type: number
    Citation:
      type: object
      required:
        - id
        - source
        - author
        - title
        - locations
      properties:
        id:
          type: string
        source:
          type: object
          required:
            - bookId
            - footnoteIdentifier
            - footnotePage
          bookId:
            type: string
          footnoteIdentifier:
            type: string
          footnotePage:
            type: string
          sourcePageText:
            type: string
        author:
          type: string
        title:
          type: string
        locationsCited:
          type: array
          items:
            type: array
            items: 
              type: object
              properties:
                type:              
                  type: string
                  enum:
                    - page
                    - chapter
                    - book
                    - volume
                    - question
                    - article
                    - lecture
                    - position
                    - verse
                  rawlabel:
                    type: string
                  value:
                    type: number
              

        